requirements.txt
fastapi
uvicorn
jinja2
psycopg2-binary
alembic
sqlalchemy
asyncpg
pydantic
bcrypt
python-jose
psycopg